+++ src/lib/editor/engine/styles/numberUnit.ts
@@ -1,22 +1,17 @@
-export class NumberUnit {
-    stringToParsedValue = (val: string, percent: boolean = false): [number, string] => {
-        const matches = val.match(/([-+]?[0-9]*\.?[0-9]+)([a-zA-Z%]*)/);
+export function stringToParsedValue(val: string, percent: boolean = false): [number, string] {
+    const matches = val.match(/([-+]?[0-9]*\.?[0-9]+)([a-zA-Z%]*)/);
 
-        let num = matches ? parseFloat(matches[1]) : 0;
-        let unit = matches && matches[2] ? matches[2] : "";
+    let num = matches ? parseFloat(matches[1]) : 0;
+    let unit = matches && matches[2] ? matches[2] : "";
 
-        // Handle percentage being fractions, if no parsed unit, convert to percentage
-        if (percent && unit === "") {
-            unit = "%";
-            num = num <= 1 ? num * 100 : num;
-        }
-        return [num, unit];
-    };
+    // Handle percentage being fractions, if no parsed unit, convert to percentage
+    if (percent && unit === "") {
+        unit = "%";
+        num = num <= 1 ? num * 100 : num;
+    }
+    return [num, unit];
+};
 
-    parsedValueToString = (
-        floatValue: number | string,
-        unit: string,
-    ): string => {
-        return `${floatValue}${unit}`;
-    };
-}
+export function parsedValueToString(floatValue: number | string, unit: string): string {
+    return `${floatValue}${unit}`;
+};
